pyproject.toml
[project]
name = "autochain-b2b-supply-chain-copilot"
version = "0.1.0"
description = ""
authors = [
    {name = "girishdigge",email = "girishdigge@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "portia-sdk-python [openai,google](>=0.6.2,<0.7.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "stripe (>=12.4.0,<13.0.0)",
    "fastapi (>=0.116.1,<0.117.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
